# problem5/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BE API",
    "version": "1.0.0",
    "description": "Backend API - Auto-generated from JSDoc"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development"
    }
  ],
  "paths": {
    "/api/cards": {
      "get": {
        "tags": [
          "Cards"
        ],
        "summary": "Get all cards",
        "responses": {
          "200": {
            "description": "List of all cards",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Card"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/cards/{id}": {
      "get": {
        "tags": [
          "Cards"
        ],
        "summary": "Get card by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Card found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          },
          "404": {
            "description": "Card not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Cards"
        ],
        "summary": "Update a card",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "cover": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "stackId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Card updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          },
          "404": {
            "description": "Card not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Cards"
        ],
        "summary": "Delete a card",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Card deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Card not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/stacks/{stackId}/cards": {
      "get": {
        "tags": [
          "Cards"
        ],
        "summary": "Get all cards in a stack",
        "parameters": [
          {
            "name": "stackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of cards",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Card"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Stack not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cards"
        ],
        "summary": "Create a card in a stack",
        "parameters": [
          {
            "name": "stackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "cover"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "cover": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Card created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          },
          "404": {
            "description": "Stack not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/cards/{id}/move": {
      "patch": {
        "tags": [
          "Cards"
        ],
        "summary": "Move a card to a different stack",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "stackId"
                ],
                "properties": {
                  "stackId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Card moved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          },
          "404": {
            "description": "Card or stack not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": [
          "General"
        ],
        "summary": "Welcome",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "General"
        ],
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/stacks": {
      "get": {
        "tags": [
          "Stacks"
        ],
        "summary": "Get all stacks with card counts",
        "responses": {
          "200": {
            "description": "List of stacks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Stack"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Stacks"
        ],
        "summary": "Create a new stack",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "cover": {
                    "$ref": "#/components/schemas/Cover"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Stack created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stack"
                }
              }
            }
          }
        }
      }
    },
    "/api/stacks/{id}": {
      "get": {
        "tags": [
          "Stacks"
        ],
        "summary": "Get stack by ID with its cards",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stack with cards",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StackWithCards"
                }
              }
            }
          },
          "404": {
            "description": "Stack not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Stacks"
        ],
        "summary": "Update a stack",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "cover": {
                    "$ref": "#/components/schemas/Cover"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stack updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stack"
                }
              }
            }
          },
          "404": {
            "description": "Stack not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Stacks"
        ],
        "summary": "Delete a stack and all its cards",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stack deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Stack not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Card": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "stackId": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "cover": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "Cover": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "color",
              "gradient",
              "image"
            ]
          },
          "value": {
            "type": "string"
          }
        }
      },
      "Stack": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "cover": {
            "$ref": "#/components/schemas/Cover"
          },
          "cardCount": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string"
          }
        }
      },
      "StackWithCards": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "cover": {
            "$ref": "#/components/schemas/Cover"
          },
          "cards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Card"
            }
          },
          "createdAt": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": []
}